+++ src/paramset_collection_construct.c
@@ -467,7 +467,7 @@ static SEXP character_vector(const char *const *values, R_xlen_t size) {
   return result;
 }
 
-static void set_table_attributes(SEXP table,
+static SEXP set_table_attributes(SEXP table,
     const char *const *column_names, R_xlen_t column_count,
     R_xlen_t row_count, int include_row_names, int sorted_by_id,
     int include_selfref) {
@@ -494,12 +494,17 @@ static void set_table_attributes(SEXP table,
     ++protected_count;
     Rf_setAttrib(table, Rf_install("sorted"), sorted);
   }
+  SEXP result = table;
   if (include_selfref) {
-    paradox_set_data_table_selfref(table);
-    Rf_setAttrib(table, R_NamesSymbol, R_NilValue);
-    Rf_setAttrib(table, R_NamesSymbol, names);
+    result = PROTECT(paradox_prepare_data_table(table));
+    ++protected_count;
+    SEXP result_names = PROTECT(Rf_getAttrib(result, R_NamesSymbol));
+    ++protected_count;
+    Rf_setAttrib(result, R_NamesSymbol, R_NilValue);
+    Rf_setAttrib(result, R_NamesSymbol, result_names);
   }
   UNPROTECT(protected_count);
+  return result;
 }
 
 static SEXP allocate_columns(const SEXPTYPE *types, R_xlen_t column_count,
@@ -518,7 +523,7 @@ static SEXP new_table(const char *const *column_names,
     const SEXPTYPE *types, R_xlen_t column_count, R_xlen_t row_count,
     int include_row_names, int sorted_by_id, int include_selfref) {
   SEXP table = PROTECT(allocate_columns(types, column_count, row_count));
-  set_table_attributes(
+  SEXP result = PROTECT(set_table_attributes(
     table,
     column_names,
     column_count,
@@ -526,9 +531,9 @@ static SEXP new_table(const char *const *column_names,
     include_row_names,
     sorted_by_id,
     include_selfref
-  );
-  UNPROTECT(1);
-  return table;
+  ));
+  UNPROTECT(2);
+  return result;
 }
 
 static void copy_at(SEXP destination, R_xlen_t destination_row,